Rework connection management to handle retries.
When an error occurs on the connection, the library will automatically
try to reconnect (within a limit of 5 retries) and fully reinitiate the
connection to the server to try and recover. A new LWM2M registration is
performed upon successful reconnection in case the server lost track of
previous registrations.

This handles the following cases:
 - The client connection goes down for a few seconds
 - The remote server lost connectivity
 - The remote server restarted

Note: for TCP connections the local port is increased by one before
each reconnection. This is needed to avait getting busy error on socket
connection due to the TCP "TIME-WAIT" mechanism. If a fixed local port
is passed to the library, one must not assume that this port will remain
valid for the whole process duration.