openssl: support arm/arm64 cross compilation
When we build this on arm/arm64 cross compilation environment,
the config script of the openssl cannot recognize the build
architecture. Thus, this patch specifies the build architecture
directly for cross compilation.
For arm64 architecture, there is an issue[1] during build. The
eventfd systcall is not defined for arm64 syscall table, so we
have to disable afalg for the arm64 architecture.

[1]: https://github.com/openssl/openssl/issues/1685

Change-Id: I91f2cc2f1fceb76ca8129f28e07228473603bdb9
Signed-off-by: Chanho Park <chanho61.park@samsung.com>